* commit 'e591553e2ab553d8301b7172be1c2b7423d76c96':
  add second moment shower shape variable to cluster ntupletools
  fix bug in ECLVariables related to eclClusterNumberOfHadronDigits

+++ analysis/VariableManager/src/ECLVariables.cc
@@ -48,7 +48,7 @@ namespace Belle2 {
 
     double eclClusterNumberOfHadronDigits(const Particle* particle)
     {
-      int result = -999;
+      double result = -999;
       const ECLCluster* cluster = particle->getECLCluster();
       if (cluster) {
         if (eclClusterHasPulseShapeDiscrimination(particle)) {
